b: fall back to current ISA instead of "unknown".
Sometimes it happens that we call "./b dis=<prog>" and cannot determine
the ISA of <prog>. Previously, we used "unknown" in this case, which of
course never works, but just leads to an error. The more sane fallback
is therefore M3_ISA, which at least starts objdump (or whatever),
although it might not be able to interpret <prog> in some cases.