Support the RoT on hw23.
This required a couple of additional changes.

1. We now have a unified way of starting on RISC-V (both on hw23 and
gem5): we let the core start at 0 and inject a jump at that address
that jumps to the actual entry point.

2. We use a different memory layout and strategy to prepare the switch
in the RoT. The context is always stored at address 0, followed by the
environment at 0x1000. The area 0x2000 - 0x4000 is used for the stack.
We always clear the context and stack before the switch as before, and
use additional symbols to clear the data and BSS.

3. As we are not using riscv-rt anymore, we need to clear the BSS
manually, because we are just loading a blob into memory that does not
include BSS. We do not before _start by injecting a corresponding
routine (_reset) into the very beginning of every RoT stage.

4. Several places were not prepared for the special needs of the
hardware platform or its Pico RISCV-32 core.