unimux: refactoring to separate preemptable from non-preemptable version.
The non-preemptable version on the Pico core is significantly simpler
than the preemptable version on Rocket/BOOM cores. As these two
versions share almost nothing, the previous implementation was pretty
confusing and hard to maintain.

For that reason, I've now moved all stuff for the preemptable version
into the module preempt and the non-preemptable version into the module
nopreempt. Both implement a bunch of functions to abstract from their
differences.

I've also removed various unused bits of code and imports, which still
existed due to an accidentally added allow(warnings) for the whole
crate.