librustm3: turned finish hook into argument to exec.
Previously, we had this hook as a method in the Mapper trait, but it
does not really belong there. We primarily used it to lock a tile and
make regions exclusive, which has nothing to do with *how* we map an
application.

I've therefore now turned it into an argument to
ChildActivity::exec_file.