unimux: only do sidecall check in no-preempt mode.
In preempt mode, we get an interrupt and really want to receive it
during interrupt handling in order to directly switch to the app if we
received the activity-start sidecall. If we accidentally get this with
an explicit sidecalls check, we do not switch to it and therefore call
run_to_completion, which is not meant to be called in preempt mode.