Changed exclusive-region concept.
We do no longer let the TCU enforce rules regarding the exclusive
regions, but let the RoT enforces them.

Therefore, instead of having external TCU commands that are used by the
kernel and restricted accordingly by the TCU, we only allow the RoT to
manipulate the exclusive-region registers to force the kernel to ask the
RoT for adding/removing such regions.

We use a new external register called EXREG_MNG to define who has the
permissions on a specific tile to manage its exclusive regions. This
register is write-once and the RoT uses the fact that it boots first to
write itself into this register on all tiles.

This simplifies the TCU significantly compared with the old approach and
also allows the RoT to do additional checks/operations. We use that now
to zero a region on remove to ensure that no one can read secrets from
it when it is reused.

The kernel now sends two new sidecalls to the RoT tile to add/remove
exclusive regions. These are unimplemented by all other muxes.
Unfortunately, the unimux crate is the one that receives all sidecalls,
but rots is the one that manages the exclusive regions. We therefore now
have a mechanism in unimux to register handlers for these two sidecalls,
which is used by rots.