imgprocpipe: workaround for current resmng limitation.
We currently cannot give our own tile access to a shared memory region,
because when revoking the memory tile that holds the exclusive region
quota for that, our application is still running on its tile. However,
the region cannot be removed until we reset this tile.

I've therefore worked around that problem now by letting the repeater
allocate the memory tile with exregs, give the childs tile access, and
delegate the memory tile to the child. This way, the child can still
give others access to the shared memory region, but the repeater can
first reset the child's tile before it revokes the memory tile.