rot: do not use dynamic EP for hash session.
It's actually quite complicated because we delegate the EP to someone
else for configuration. So, for example, the application obtains the EP
from rots and delegates it to m3fs. Now m3fs could invalidate the EP at
any point in time or revoke the gate that is configured on the EP,
leading also to an invalidate. In both case, the dynamic bit will be
unset. But who is going to set it again? m3fs doesn't know that it's a
dynamic EP, neither does the application. Only rots does, but it is no
longer involved.

So, let's not use a dynamic EP here, but simply unfreeze it whenever
it's necessary. I think that also shows again that the dynamic bit was a
mistake. It makes things much more complicated and should make basically
no difference performance wise, I think.