b: improved argument handling.
Previously, we had the issue that "--" stopped argparse to consume
options, but not positional arguments. Commands therefore still got
their positional arguments from the arguments after "--". That's in our
case not desirable I think. Instead, we want to stop collecting
arguments for ourself after "--" altogether and collect the rest to pass
them on to tools we call.

For example, if we run "./b cargo -- --help" we want "--help" to be
passed to cargo and not see the help of b. That didn't work previously,
because argparse consumed these as well.

We are now doing the split at "--" ourself and only pass the arguments
before that to argparse. The collection "remainder" arguments from
argparse are then concatenated with the arguments after "--" as the
remainder arguments for the subcommand.

This allows us to use commands that do not expected positional arguments
with "--", but also to use "--" when we want to be sure that arguments
are passed on to, e.g., cargo.