rot: do not put context at address 0.
Using null pointers is undefined behavior in Rust and we therefore
shouldn't use them. This worked previously, because the old Rust version
(or probably LLVM?) did not handle this case. The new version simply
generates an "unimpl" instruction if it knows that it's a null pointer
and removes the remaining code.