rot: make rosa and rots fit into SPM again.
Unfortunately, the binaries are larger when using the new Rust version,
so that they did not fit into the SPM anymore. I've already reduced some
stuff before, but that didn't suffice.

The next escalation is now to let rosa and rots slightly overlap. They
do run one after the other after all. However, rosa is loading rots and
thus rots cannot completely overwrite rosa. But there are functions in
rosa of which we know that they are not used anymore. The current layout
is that rosa sits at the end of the SPM and rots is loaded to the front.
We therefore need to put functions at the beginning of rosa that can be
overwritten. This is _reset and stage1. _reset cannot be moved and
stage1 is pretty large and is definitely not used anymore after we
entered stage2.