Allow C++ apps to use unimux.
We now have a small separate rust crate called unimuxheap, which
provides the heap for unimux in case we do not link a Rust app against
it. For C++ apps we therefore link against both unimux and unimuxheap
and provide our own heap in the application, whereas the heap is shared
between unimux and Rust apps.