Simplified building of TEEs.
Previously, we required TEEs to be special by, for example, not using
the M³ standard library, defining a heap, defining env_run manually, and
so on. Only some of these things were actually required though.

I've now reworked the whole procedure, so that we can turn normal
applications into TEEs by just linking them differently. This works,
because unimux now has its own heap (although that's actually not used)
and the application has a separate one. We only have to use the unimux
startup code and link against its library. Unimux will therefore handle
the communication with the kernel at the beginning and then jump to the
application (env_run) afterwards.

There is one special case: rots shares it's heap with unimux and
therefore defines it in rots. Additionally, it's the only case where we
use the unimux crate in order to interact with it. TEE applications only
use ecalls to interact with unimux.