m3fs: support shared-memory backend.
We can now use a shared-memory area created by a resource manager for
an m3fs filesystem. This will be an empty filesystem and thus needs
special support in m3fs.

This allows us to exclusively share the filesystem between m3fs and its
clients by making this shared-memory region exclusive for the involved
tiles.